Edited the current timeline progression system
Changed forward terminus' if statement, as prior it was basing off of a sample clip and would always return true. Along with when moving to the next keyframe I just set newKeyTime to 0, as the prior method would skip ahead. Current issue is that the start of the clip is sped up for some reason.